ptr problem get_library_path()

+++ astrology_macos/Types/astrology_macos-Bridging-Header.h
@@ -16,7 +16,7 @@ char *swe_version(char *s);
 char *swe_get_library_path(char *path);
 
 // Swe03
-int swe_calc_ut(float tjd_ut, int ipl, int iflag, float *p_xx, char *p_serr);
+int swe_calc_ut(double tjd_ut, int ipl, int iflag, double *p_xx, char *p_serr);
 
 // Swe08
 double swe_julday(int year, int month, int day, double hour, int calandar);
